namada_vote_ext
Module bridge_pool_roots

Vote extension types for adding a signature of the bridge pool merkle root to be added to storage. This will be used to generate bridge pool inclusion proofs for Ethereum.

Structs

BridgePoolRootVext

A vote extension containing a validator’s signature of the current root and nonce of the Ethereum bridge pool.

MultiSignedVext

A collection of validator signatures over the Ethereum bridge pool Merkle root and nonce.

SignedVext

A signed BridgePoolRootVext.

Type Aliases

Vext

Alias for BridgePoolRootVext.