namada_vote_ext::bridge_pool_roots
Type Alias Vext

pub type Vext = BridgePoolRootVext;

Alias for BridgePoolRootVext.

Aliased Type

struct Vext {
    pub validator_addr: Address,
    pub block_height: BlockHeight,
    pub sig: CommonSignature,
}

Fields
validator_addr: Address

The address of the validator who submitted the vote extension.

block_height: BlockHeight

The block height at which the vote extensions was sent.

This can be used as replay protection as well as allowing validators to query the epoch with the appropriate validator set to verify signatures

sig: CommonSignature

The actual signature being submitted. This is a signature over keccak(eth_header || keccak(root || nonce)).

Implementations

impl Vext

pub fn sign(&self, sk: &SecretKey) -> SignedVext

Creates a new signed Vext.