namada_vote_ext
Module validator_set_update

Contains types necessary for processing validator set updates in vote extensions.

Structs

EthAddrBook

Container type for both kinds of Ethereum bridge addresses:

SerializeWithAbiEncode

Tag type that indicates we should use AbiEncode to sign data in a namada_tx::Signed wrapper.

SignedVext

Represents a Vext signed by some validator, with an Ethereum key.

ValidatorSetArgs

Struct for serializing validator set arguments with ABI for Ethereum smart contracts.

ValidatorSetUpdateVext

Represents a validator set update, for some new Epoch.

ValidatorSetUpdateVextDigest

Contains the digest of all signatures from a quorum of validators for a Vext.

Traits

VotingPowersMapExt

This trait contains additional methods for a VotingPowersMap, related with validator set update vote extensions logic.

Functions

valset_upd_toks_to_hashes

Returns the bridge and governance keccak hashes calculated from the given hot and cold key addresses, and their respective validator’s voting powers, normalized to 2^32.

Type Aliases

Vext

Type alias for a ValidatorSetUpdateVext.

VextDigest

Type alias for a ValidatorSetUpdateVextDigest.

VotingPowersMap

Provides a mapping between EthAddress and token::Amount instances.