namada_vote_ext::validator_set_update
Function valset_upd_toks_to_hashes

pub fn valset_upd_toks_to_hashes(
    next_epoch: Epoch,
    bridge_validators: Vec<Token>,
    governance_validators: Vec<Token>,
) -> (KeccakHash, KeccakHash)

Returns the bridge and governance keccak hashes calculated from the given hot and cold key addresses, and their respective validator’s voting powers, normalized to 2^32.