namada_vote_ext::validator_set_update
Trait VotingPowersMapExt

pub trait VotingPowersMapExt {
    // Required method
    fn get_sorted(&self) -> Vec<(&EthAddrBook, &Amount)>;

    // Provided methods
    fn get_abi_encoded(&self) -> (Vec<Token>, Vec<Token>) { ... }
    fn get_bridge_and_gov_hashes(
        &self,
        next_epoch: Epoch,
    ) -> (KeccakHash, KeccakHash) { ... }
}

This trait contains additional methods for a VotingPowersMap, related with validator set update vote extensions logic.

Required Methods

fn get_sorted(&self) -> Vec<(&EthAddrBook, &Amount)>

Returns a Vec of pairs of validator addresses and voting powers, sorted in descending order by voting power.

Provided Methods

fn get_abi_encoded(&self) -> (Vec<Token>, Vec<Token>)

Returns the list of Ethereum validator hot and cold addresses and their respective voting powers (in this order), with an Ethereum ABI compatible encoding. Implementations of this method must be deterministic based on self. In addition, the returned Vecs must be sorted in descending order by voting power, as this is more efficient to deal with on the Ethereum side when working out if there is enough voting power for a given validator set update.

fn get_bridge_and_gov_hashes( &self, next_epoch: Epoch, ) -> (KeccakHash, KeccakHash)

Returns the bridge and governance keccak hashes of this VotingPowersMap.