namada_vote_ext::validator_set_update
Type Alias VextDigest

pub type VextDigest = ValidatorSetUpdateVextDigest;

Type alias for a ValidatorSetUpdateVextDigest.

Aliased Type

struct VextDigest {
    pub signatures: IndexMap<Address, CommonSignature>,
    pub voting_powers: IndexMap<EthAddrBook, Amount>,
}

Fields
signatures: IndexMap<Address, CommonSignature>

A mapping from a consensus validator address to a Signature.

voting_powers: IndexMap<EthAddrBook, Amount>

The addresses of the validators in the new Epoch, and their respective voting power.

Implementations

impl VextDigest

pub fn singleton(SignedVext: SignedVext) -> VextDigest

Build a singleton VextDigest, from the provided Vext.

pub fn decompress(self, signing_epoch: Epoch) -> Vec<SignedVext>

Decompresses a set of signed Vext instances.