Crate namada_vote_ext

This module contains types necessary for processing vote extensions.

Modules

bridge_pool_roots

Vote extension types for adding a signature of the bridge pool merkle root to be added to storage. This will be used to generate bridge pool inclusion proofs for Ethereum.

ethereum_events

Contains types necessary for processing Ethereum events in vote extensions.

ethereum_tx_data_variants

All the variants of EthereumTxData, stored in a module.

validator_set_update

Contains types necessary for processing validator set updates in vote extensions.

Structs

VoteExtension

This type represents the data we pass to the extension of a vote at the PreCommit phase of Tendermint.

Enums

EthereumTxData

Data associated with Ethereum protocol transactions.

Traits

EthereumTxDataVariants

All the variants of EthereumTxData, stored in a trait.